requirements.txt
flask==2.3.3
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
chromadb==0.4.15
langchain==0.0.267
langchain-community==0.0.8
pdf2image==1.16.3
pytesseract==0.3.10
pypdf2==3.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
uuid==1.30
numpy==1.25.2
tqdm==4.66.1
aiohttp==3.8.5
requests==2.31.0
cryptography==45.0.3